Added a way to edit
Adding a way to instead of sending a whole new message, editing an existing one must be useful.
You just need to do  
```
paginationEmbed(msg, pages, ['⏪', '⏩'], 60000, true)
``` 
adding the true will make it edit a already sent message instead of creating a new one. (The option as default is false
In my opinion this should be useful.